+++ Packages/com.unity.cloud.gltfast/CHANGELOG.md
@@ -7,12 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- `LogCode.AnimationComponentFail`
 
 ### Changed
 - Clarified [IDeferAgent.ShouldDefer](xref:GLTFast.IDeferAgent.ShouldDefer) documentation to note that it must eventually return `false`, otherwise imports may stall indefinitely without raising an error.
 
 ### Fixed
 - Fixed false positives in export to stream tests because it actually validated results from non-stream tests.
+- Prevent exception when Animation component was not created successfully.
 
 ### Removed